# product-api/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    description: GenericError GenericError generic error
    properties:
      message:
        description: message
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/rabadiyaronak/microservice-go/product-api/sdk/models
  Product:
    description: Product defines model of product API
    properties:
      description:
        description: the description for this poduct
        maxLength: 10000
        type: string
        x-go-name: Description
      id:
        description: id for the product
        format: int64
        minimum: 1
        pattern: '[0-9]+'
        type: integer
        x-go-name: Id
      name:
        description: name for the product
        maxLength: 255
        type: string
        x-go-name: Name
      price:
        description: the price for the product
        format: float
        minimum: 0.01
        type: number
        x-go-name: Price
      sku:
        description: the SKU for the product
        pattern: '[a-z]+-[a-z]+-[a-z]+'
        type: string
        x-go-name: SKU
    required:
    - name
    - price
    - sku
    type: object
    x-go-package: github.com/rabadiyaronak/microservice-go/product-api/data
  ValidationError:
    description: ValidationError ValidationError validation error
    properties:
      message:
        description: message
        items:
          type: string
        type: array
        x-go-name: Message
    type: object
    x-go-package: github.com/rabadiyaronak/microservice-go/product-api/sdk/models
host: localhost:9090
info:
  description: Documantation for Product Api
  title: Prodcut Api
  version: 1.0.0
paths:
  /products:
    get:
      description: Returns a list of products
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
    post:
      description: Create a new product
      operationId: createProduct
      parameters:
      - description: |-
          Product data structure to Update or Create.
          Note: the id field is ignored by update and create operations
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Product'
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "400":
          $ref: '#/responses/errorResponse'
        "422":
          $ref: '#/responses/errorValidation'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
  /products/{id}:
    delete:
      description: Update a products details
      operationId: deleteProduct
      parameters:
      - format: int64
        in: path
        minimum: 1
        name: id
        pattern: '[0-9]+'
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
    get:
      description: Returns Product with given id
      operationId: getProductById
      parameters:
      - format: int64
        in: path
        minimum: 1
        name: id
        pattern: '[0-9]+'
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
    put:
      description: Update a products details
      operationId: updateProduct
      parameters:
      - format: int64
        in: path
        minimum: 1
        name: id
        pattern: '[0-9]+'
        required: true
        type: integer
        x-go-name: ID
      - description: |-
          Product data structure to Update or Create.
          Note: the id field is ignored by update and create operations
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Product'
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "422":
          $ref: '#/responses/errorValidation'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  errorValidation:
    description: Validation errors defined as an array of strings
    schema:
      $ref: '#/definitions/ValidationError'
  noContentResponse:
    description: No content is returned by this API endpoint
  productResponse:
    description: Wrapper ds to represent a single product
    schema:
      $ref: '#/definitions/Product'
  productsResponse:
    description: A list of products returns in the response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
swagger: "2.0"